+++ src/App.tsx
@@ -1,8 +1,8 @@
-import './index.css';
-import { CartPage } from './features/Cart/pages/CartPage';
-import { AppLayout } from './shared/components/AppLayout/AppLayout';
-import { ToastProvider } from './shared/context/ToastProvider';
-import { CartProvider } from './features/Cart/context/CartProvider';
+import '@/index.css';
+import { CartPage } from '@/features/Cart/pages/CartPage';
+import { AppLayout } from '@/shared/components/AppLayout/AppLayout';
+import { ToastProvider } from '@/shared/context/ToastProvider';
+import { CartProvider } from '@/features/Cart/context/CartProvider';
 
 export const App = () => {
   return (
