+++ src/api/fetcher.ts
@@ -1,5 +1,5 @@
 import { buildQueryParams } from '@/shared/utils/url';
-import { ENV } from './env';
+import { ENV } from '@/api/env';
 import { HttpError } from '@/shared/errors/HttpError';
 
 type FetcherOptions<T> = {
