+++ src/shared/components/AppLayout/AppLayout.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react';
 
-import { StyledAppLayout } from './AppLayout.styled';
+import { StyledAppLayout } from '@/shared/components/AppLayout/AppLayout.styled';
 
 export const AppLayout = ({ children }: PropsWithChildren) => {
   return <StyledAppLayout>{children}</StyledAppLayout>;
