+++ src/shared/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import { ComponentProps, ReactNode } from 'react';
 
-import { StyledHeader } from './Header.styled';
+import { StyledHeader } from '@/shared/components/Header/Header.styled';
 
 export type HeaderProps = {
   left: ReactNode;
